Create project and add dependencies
This project is a test provide by Xebia. The main theme is Harry Potter universe, more particulary with the library. The main goal is to display a list pf books and allow the user to select them. Each selection has to be added in a cart. A proposal will be computed depending the user selection.