Add recyclerView and perform some improvement
The first screen is set, we have a list of of books that are displaying with RecyclerView.
An adapter and viewHolder are configured and the data are fetch with retrofit threw the viewModel which is injected by dagger. Now the next stage is to create another screen which show the cart recap.