Add commercial offer features
A second Api is called with retrofit, the use of data binding and view model is the same process. The difference it's we are not getting a list of object but an object which allow us to perform computations. The main idea is to select an amount of product, to store them and to show in the cart icon the quantity. When user perform a click on the cart icon a new activity with the best proposal will be displayed.