Handle data provide by the api
The idea here is to work with the data in order to handle the commercial offer. The first things was to communicate between activities and to store data inside the view model of each Ui. There are two apis so to call are necessary to compute the data. The next step is to compare the bests offers between three options.