Add badge on icon menu
This technique allow to draw a circle and a text inside of it to display the number of item added inside the cart, using the layer drawable from layer list xml. This avoid to transform a layout into a drawable or to store x number of image to represent each case.